[FIX] 버튼 만드는 로직 buttonRenderer로 옮김
예잉
