+++ js/utils/uiUtils/buttonManager/employee/compensationPlanning/home.js
@@ -2,10 +2,6 @@ import { BUTTON } from '../../../../../../config/employee/compensationPlanning/c
 import { viewInformationList } from '../../../tableRenderer/employee/compensationPlanning/informationList.js';
 import { informationType } from '../../../tableRenderer/employee/compensationPlanning/informationList.js';
 
-export const renderButtons = () => {
-  initialButtons(BUTTON.TASK.COMPENSATIONPLANNING.HOME, compensationPlanningTaskMapper);
-}
-
 const viewEvaluatePartnerCompany = () => {
   viewInformationList(informationType.EVALUATE_PARTNERCOMPANY);
 }
@@ -14,9 +10,32 @@ const viewManagementPartnerCompany = () => {
   viewInformationList(informationType.MANAGEMENT_PARTNERCOMPANY);
 }
 
+const evaluate = () => {
+  alert("평가-보상기획");
+}
+
+const cancel = () => {
+  window.location.href = "informationList.html";
+}
+
 const compensationPlanningTaskMapper = {
-  EVALUATE_PARTNERCOMPANY: viewEvaluatePartnerCompany,
-  MANAGEMENT_PARTNERCOMPANY: viewManagementPartnerCompany
+  HOME: {
+    EVALUATE_PARTNERCOMPANY: viewEvaluatePartnerCompany,
+    MANAGEMENT_PARTNERCOMPANY: viewManagementPartnerCompany
+  },
+  EVALUATE_PARTNERCOMPANY: {
+    EVALUATE: evaluate,
+    CANCEL: cancel
+  }
+}
+
+export const renderButtons = () => {
+  initialButtons(BUTTON.TASK.COMPENSATIONPLANNING.HOME, compensationPlanningTaskMapper.HOME);
+}
+
+export const renderDetailButtons = () => {
+  const type = sessionStorage.getItem("currentType");
+  initialButtons(BUTTON.TASK.COMPENSATIONPLANNING[type], compensationPlanningTaskMapper[type]);
 }
 
 const initialButtons = (buttonMessages, buttonActionMapper) => {
