+++ js/utils/uiUtils/buttonManager/employee/contractManagement/home.js
@@ -2,10 +2,6 @@ import { BUTTON } from '../../../../../../config/employee/contractManagement/con
 import { viewInformationList } from '../../../tableRenderer/employee/contractManagement/informationList.js';
 import { informationType } from '../../../tableRenderer/employee/contractManagement/informationList.js';
 
-export const renderButtons = () => {
-  initialButtons(BUTTON.TASK.CONTRACT_MANAGEMENT.HOME, contractManagementTaskMapper);
-};
-
 const viewDefaultContract = () => {
   viewInformationList(informationType.DEFAULT_CONTRACT);
 }
@@ -26,12 +22,71 @@ const viewTermination = () => {
   viewInformationList(informationType.TERMINATION);
 }
 
+const sendNotice = () => {
+  alert("안내장 발송");
+}
+const reviewRecontract = () => {
+  alert("재계약 승인");
+}
+const denyRecontract = () => {
+  alert("재계약 거절");
+}
+const reviewEndorsement = () => {
+  alert("배서 승인");
+}
+const denyEndorsement = () => {
+  alert("배서 거절");
+}
+const reviewRevival = () => {
+  alert("부활 승인");
+}
+const denyRevival = () => {
+  alert("부활 거절");
+}
+const requestTerminationFee = () => {
+  alert("해지 승인");
+}
+const cancel = () => {
+  window.location.href = "informationList.html";
+}
+
 const contractManagementTaskMapper = {
-  DEFAULT_CONTRACT: viewDefaultContract,
-  RECONTRACT: viewRecontract,
-  ENDORSEMENT: viewEndorsement,
-  REVIVAL: viewRevival,
-  TERMINATION: viewTermination
+  HOME: {
+    DEFAULT_CONTRACT: viewDefaultContract,
+    RECONTRACT: viewRecontract,
+    ENDORSEMENT: viewEndorsement,
+    REVIVAL: viewRevival,
+    TERMINATION: viewTermination
+  },
+  DEFAULT_CONTRACT: {
+    NOTICE: sendNotice,
+    CANCEL: cancel
+  },
+  RECONTRACT: {
+    APPLY: reviewRecontract,
+    DENY: denyRecontract
+  },
+  ENDORSEMENT: {
+    APPLY: reviewEndorsement,
+    DENY: denyEndorsement
+  },
+  REVIVAL: {
+    APPLY: reviewRevival,
+    DENY: denyRevival
+  },
+  TERMINATION: {
+    REQUEST: requestTerminationFee,
+    CANCEL: cancel
+  }
+}
+
+export const renderButtons = () => {
+  initialButtons(BUTTON.TASK.CONTRACT_MANAGEMENT.HOME, contractManagementTaskMapper.HOME);
+};
+
+export const renderDetailButtons = () => {
+  const type = sessionStorage.getItem("currentType");
+  initialButtons(BUTTON.TASK.CONTRACT_MANAGEMENT[type], contractManagementTaskMapper[type]);
 }
 
 const initialButtons = (buttonMessages, buttonActionMapper) => {
