+++ js/utils/uiUtils/tableRenderer/employee/compensation/detail.js
@@ -3,7 +3,7 @@ import {
   fetchGetReportById
 } from "../../../../apiUtils/apiDocumentation/employee/compensation/compensation.js";
 import {BUTTON} from "../../../../../../config/employee/compensation/compensation.js";
-import {BUTTON as COMMON_BUTTON} from "../../../../../../config/common.js";
+import {renderDetailButtons} from "../../../buttonManager/employee/compensation/home.js";
 
 const serviceTypeStr = {
   EmergencyTowing: "긴급견인",
@@ -78,7 +78,7 @@ export const renderDetails = async () => {
   if (selectedDataId) {
     const selectedData = await context[type].fetchGetById(selectedDataId);
     renderDetailsTable(selectedData);
-    renderButtons();
+    renderDetailButtons();
   }
 }
 
@@ -126,47 +126,4 @@ const renderDetailsTable = (data) => {
     }
     detailsTable.querySelector("tbody").appendChild(row);
   });
-}
-
-const renderButtons = () => {
-  initialButtons();
-}
-
-const initialButtons = () => {
-  const buttonContainer = document.getElementById("buttonContainer");
-  const type = sessionStorage.getItem("currentType");
-  const buttonMessages = context[type].buttons;
-  // 객체의 각 항목을 순회하여 버튼 생성
-  Object.entries(buttonMessages).forEach(([key, name]) => {
-    const button = document.createElement("div");
-    button.className = "button-item";
-    button.textContent = name; // 버튼에 표시할 텍스트 설정
-
-    button.addEventListener("click", compensationTaskMapper[type][key]);
-
-    buttonContainer.appendChild(button); // 버튼을 buttonContainer에 추가
-  });
-}
-
-const requestCompensation = () => {
-  alert("보상처리");
-}
-
-const requestInsuranceMoney = () => {
-  alert("보험금 지급");
-}
-
-const cancel = () => {
-  window.location.href = "informationList.html";
-}
-
-const compensationTaskMapper = {
-  REQUEST_COMPENSATION: {
-    REQUEST: requestCompensation,
-    CANCEL: cancel
-  },
-  REQUEST_INSURANCE_MONEY: {
-    REQUEST: requestInsuranceMoney,
-    CANCEL: cancel
-  }
 }
