+++ js/utils/uiUtils/tableRenderer/employee/contractManagement/detail.js
@@ -5,7 +5,7 @@ import {
   fetchGetRevivalById,
   fetchGetTerminationById
 } from "../../../../apiUtils/apiDocumentation/employee/contractManagement/contractManagement.js";
-import {BUTTON} from "../../../../../../config/employee/contractManagement/contractManagement.js";
+import {renderDetailButtons} from "../../../buttonManager/employee/contractManagement/home.js";
 
 const accidentHistoryDetail = (data) => {
   return [
@@ -94,28 +94,23 @@ const terminationDetail = (data) => {
 const context = {
   DEFAULT_CONTRACT: {
     detailGetter: contractDetail,
-    fetchGetById: fetchGetContractById,
-    buttons: BUTTON.TASK.CONTRACT_MANAGEMENT.DEFAULT_CONTRACT
+    fetchGetById: fetchGetContractById
   },
   RECONTRACT: {
     detailGetter: recontractDetail,
-    fetchGetById: fetchGetReContractById,
-    buttons: BUTTON.TASK.CONTRACT_MANAGEMENT.RECONTRACT
+    fetchGetById: fetchGetReContractById
   },
   ENDORSEMENT: {
     detailGetter: endorsementDetail,
-    fetchGetById: fetchGetEndorsementById,
-    buttons: BUTTON.TASK.CONTRACT_MANAGEMENT.ENDORSEMENT
+    fetchGetById: fetchGetEndorsementById
   },
   REVIVAL: {
     detailGetter: revivalDetail,
-    fetchGetById: fetchGetRevivalById,
-    buttons: BUTTON.TASK.CONTRACT_MANAGEMENT.REVIVAL
+    fetchGetById: fetchGetRevivalById
   },
   TERMINATION: {
     detailGetter: terminationDetail,
-    fetchGetById: fetchGetTerminationById,
-    buttons: BUTTON.TASK.CONTRACT_MANAGEMENT.TERMINATION
+    fetchGetById: fetchGetTerminationById
   }
 }
 
@@ -128,7 +123,7 @@ export const renderDetails = async () => {
   if (selectedDataId) {
     const selectedData = await context[type].fetchGetById(selectedDataId);
     renderDetailsTable(selectedData);
-    renderButtons();
+    renderDetailButtons();
   }
 }
 
@@ -176,75 +171,4 @@ const renderDetailsTable = (data) => {
     }
     detailsTable.querySelector("tbody").appendChild(row);
   });
-}
-
-const renderButtons = () => {
-  initialButtons();
-}
-
-const initialButtons = () => {
-  const buttonContainer = document.getElementById("buttonContainer");
-  const type = sessionStorage.getItem("currentType");
-  const buttonMessages = context[type].buttons;
-  // 객체의 각 항목을 순회하여 버튼 생성
-  Object.entries(buttonMessages).forEach(([key, name]) => {
-    const button = document.createElement("div");
-    button.className = "button-item";
-    button.textContent = name; // 버튼에 표시할 텍스트 설정
-
-    button.addEventListener("click", contractManagementTaskMapper[type][key]);
-
-    buttonContainer.appendChild(button); // 버튼을 buttonContainer에 추가
-  });
-}
-
-const sendNotice = () => {
-  alert("안내장 발송");
-}
-const reviewRecontract = () => {
-  alert("재계약 승인");
-}
-const denyRecontract = () => {
-  alert("재계약 거절");
-}
-const reviewEndorsement = () => {
-  alert("배서 승인");
-}
-const denyEndorsement = () => {
-  alert("배서 거절");
-}
-const reviewRevival = () => {
-  alert("부활 승인");
-}
-const denyRevival = () => {
-  alert("부활 거절");
-}
-const requestTerminationFee = () => {
-  alert("해지 승인");
-}
-const cancel = () => {
-  window.location.href = "informationList.html";
-}
-
-const contractManagementTaskMapper = {
-  DEFAULT_CONTRACT: {
-    NOTICE: sendNotice,
-    CANCEL: cancel
-  },
-  RECONTRACT: {
-    APPLY: reviewRecontract,
-    DENY: denyRecontract
-  },
-  ENDORSEMENT: {
-    APPLY: reviewEndorsement,
-    DENY: denyEndorsement
-  },
-  REVIVAL: {
-    APPLY: reviewRevival,
-    DENY: denyRevival
-  },
-  TERMINATION: {
-    REQUEST: requestTerminationFee,
-    CANCEL: cancel
-  }
 }
